DOC : schedule #20 move generateColor function to util
