+++ src/utils/common/generateColor.ts
