+++ src/utils/index.ts
@@ -1,2 +1,5 @@
-export { isNameValid } from './team/isNameValid';
+// common
 export { onClickOutside } from './common/onClickOutside';
+export { generateColor } from './common/generateColor';
+// team
+export { isNameValid } from './team/isNameValid';
